[project-vvm-async-api] `$OUT_DIR`を使うものをtest_utilクレートに移動 (#515)
`$OUT_DIR`を使うものをtest_utilクレートに移動
