+++ Cargo.toml
@@ -15,6 +15,7 @@ publish = false
 
 [workspace.dependencies]
 anyhow = "1.0.65"
+async_zip = { version = "0.0.11", features = ["full"] }
 clap = { version = "4.0.10", features = ["derive"] }
 const-default = { version = "1.0.0", features = ["derive"] }
 easy-ext = "1.0.1"
