+++ crates/test_util/Cargo.toml
@@ -4,6 +4,11 @@ version = "0.0.0"
 edition.workspace = true
 publish.workspace = true
 
+[dependencies]
+async_zip.workspace = true
+once_cell.workspace = true
+tokio.workspace = true
+
 [build-dependencies]
 anyhow.workspace = true
 async-std = { version = "1.12.0", features = ["attributes"] }
